requirements.txt
alembic==1.5.5
appnope==0.1.2
autopep8==1.5.5
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cheroot==8.5.2
click==7.1.2
cryptography==3.4.6
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Compress==1.9.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gssapi==1.6.12
html5lib==1.1
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.functools==3.2.1
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
ldap3==2.9
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
more-itertools==8.7.0
numpy==1.20.1
pandas==1.2.3
paramiko==2.7.2
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pgadmin4==5.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.0
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
speaklater==1.3
speaklater3==1.4
SQLAlchemy==1.3.23
sqlparse==0.4.1
sshtunnel==0.4.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
